--- a/esphome/components/ili9xxx/ili9xxx_display.h
+++ b/esphome/components/ili9xxx/ili9xxx_display.h
@@ -54,6 +54,10 @@ class ILI9XXXDisplay : public PollingComponent,
   void invert_display_(bool invert);
   void reset_();
 
+  // `buffer_2_` is an optional FB containing the currently displayed screen content. Used to determine screen update to
+  // be pushed to display.
+  uint8_t *buffer_2_{nullptr};
+
   int16_t width_{0};   ///< Display width as modified by current rotation
   int16_t height_{0};  ///< Display height as modified by current rotation
   uint16_t x_low_{0};
@@ -75,7 +79,7 @@ class ILI9XXXDisplay : public PollingComponent,
 
   uint16_t transfer_buffer_[ILI9XXX_TRANSFER_BUFFER_SIZE];
 
-  uint32_t buffer_to_transfer_(uint32_t pos, uint32_t sz);
+  uint32_t buffer_to_transfer_(uint32_t pos, uint32_t pos_16, uint32_t sz);
 
   GPIOPin *reset_pin_{nullptr};
   GPIOPin *dc_pin_{nullptr};
@@ -84,6 +88,7 @@ class ILI9XXXDisplay : public PollingComponent,
   bool prossing_update_ = false;
   bool need_update_ = false;
   bool is_18bitdisplay_ = false;
+  bool first_update_ = true;
 };
 
 //-----------   M5Stack display --------------