# What does this implement/fix?

<!-- Quick description and explanation of changes -->
Improve transfer speed to SPI display - example WT32-SC01 480x320px display 16 bit color from ~215ms to 105ms.

Use improvement to add a second framebuffer so full display redraw worst case takes ~170ms. Shortcut for full screen refresh is ~115ms.

Second framebuffer is optional and used to detect changes by comparing last display to current display. Goal is to make writing of display lambda functions easier. As I can now use `fill` operation etc.

Typical ESP32 devices with display have PSRam. Smallest possible PSRam is 2MB. One framebuffer with 480x320px and 16 bit color uses 300kb. So there should always be enough memory for second framebuffer.

All performance gain only on ESP32 Arduino. ESP32 IDF SPI is very very slow.

## Types of changes

- [ ] Bugfix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Other

**Related issue or feature (if applicable):** fixes <link to issue>

**Pull request in [esphome-docs](https://github.com/esphome/esphome-docs) with documentation (if applicable):** esphome/esphome-docs#<esphome-docs PR number goes here>

## Test Environment

- [x] ESP32
- [x] ESP32 IDF
- [ ] ESP8266
- [ ] RP2040

## Example entry for `config.yaml`:
<!--
  Supplying a configuration snippet, makes it easier for a maintainer to test
  your PR. Furthermore, for new integrations, it gives an impression of how
  the configuration would look like.
  Note: Remove this section if this PR does not have an example entry.
-->

```yaml
# Example config.yaml

```

## Checklist:
  - [x] The code change is tested and works locally.
  - [ ] Tests have been added to verify that the new code works (under `tests/` folder).

If user exposed functionality or configuration variables are added/changed:
  - [ ] Documentation added/updated in [esphome-docs](https://github.com/esphome/esphome-docs).